requirements.txt
alembic==1.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.10.0
backoff==2.2.1
boto3==1.40.20
botocore==1.40.20
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
coloredlogs==14.0
dagit==1.11.8
dagster==1.11.8
dagster-graphql==1.11.8
dagster-pipes==1.11.8
dagster-webserver==1.11.8
dagster_shared==1.11.8
docstring_parser==0.17.0
filelock==3.19.1
fsspec==2025.7.0
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.74.0
grpcio-health-checking==1.74.0
h11==0.16.0
httptools==0.6.4
humanfriendly==10.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
mysql-connector-python==9.4.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pluggy==1.6.0
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
structlog==25.4.0
tabulate==0.9.0
tomli==2.2.1
tomlkit==0.13.3
toposort==1.10
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
universal_pathlib==0.2.6
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchdog==5.0.3
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
